backtesting/requirements.txt
# Backtesting Module Requirements
psycopg2-binary>=2.9.0
numpy>=1.21.0
pyyaml>=6.0
ccxt>=4.0.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
